fixed and filled the database with movies

+++ MyMovies/MyMovies/ApplicationContext.cs
@@ -10,18 +10,29 @@ namespace MyMovies
     class ApplicationContext: DbContext
     {
         public DbSet<Movie> Movies { get; set; }
-        public DbSet<Person> Persons { get; set; }
+        public DbSet<Actor> Actors { get; set; }
+        public DbSet<Director> Directors { get; set; }
         public DbSet<Tag> Tags { get; set; }
 
-        public ApplicationContext()
+        public ApplicationContext(bool flag)
         {
-            Database.EnsureDeleted();
+            if (flag)
+            {
+                Database.EnsureDeleted();
+            }
             Database.EnsureCreated();
         }
 
         protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
         {
-            optionsBuilder.EnableSensitiveDataLogging().UseSqlServer(@"Server=localhost\SQLEXPRESS;Database=Movies;Trusted_Connection=True;");
+            optionsBuilder.UseSqlServer(@"Server=localhost\SQLEXPRESS;Database=Movies;Trusted_Connection=True;");
+        }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            modelBuilder.Entity<Movie>().HasMany(movie => movie.Actors).WithMany(actor => actor.Movies);
+            modelBuilder.Entity<Movie>().HasOne(movie => movie.Director).WithMany(director => director.Movies);
+            modelBuilder.Entity<Movie>().HasMany(movie => movie.Tags).WithMany(tag => tag.Movies);
         }
     }
 }
